requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
google-adk>=0.1.0
PyMuPDF>=1.23.0
hypothesis>=6.92.0
pytest>=7.4.0
pandas>=2.1.0
reportlab>=4.0.0
